+++ capstone/src/main/java/com/techelevator/view/Drink.java
@@ -1,8 +1,10 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product {
 
-    public Drink(String name, double price, String slot, int inventory) {
+    public Drink(String name, BigDecimal price, String slot, int inventory) {
         super(name, price, slot, inventory);
 
     }
